Improve the StatsD aspect interface
If either alias or block is given, the default key is not used. This is because
when a block is given it's easier to asume that the block will contain the
complete key.

There is no magic between default values, alias values and block values.

Alias and block are concatenated, forcing the user to include a `'.'` at the end
of the alias value making it explicit that they are concatenated.